requirements/base.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
python-decouple==3.6
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.11.0